+++ solstice/src/main/java/dev/equo/ide/BuildPluginIdeMain.java
@@ -24,6 +24,7 @@ import dev.equo.solstice.p2.WorkspaceRegistry;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.LinkedHashMap;
@@ -88,11 +89,20 @@ public class BuildPluginIdeMain {
 			Objects.requireNonNull(showConsoleFlag);
 			Objects.requireNonNull(cleanFlag);
 
-			var classpathSorted = Launcher.copyAndSortClasspath(classpath);
+			ArrayList<File> classpathSorted = Launcher.copyAndSortClasspath(classpath);
 			SignedJars.stripIfNecessary(classpathSorted);
 			var nestedJarFolder = new File(workspaceDir, NestedJars.DIR);
 			for (var nested : NestedJars.inFiles(classpathSorted).extractAllNestedJars(nestedJarFolder)) {
 				classpathSorted.add(nested.getValue());
+			}
+			if (useAtomos) {
+				// for Eclipse 4.27, we have patched the EquinoxBundle class so that it handles
+				// `getEntry("/")`
+				var version = Patch.detectVersion(classpathSorted, "org.eclipse.osgi");
+				if ("3.18.300".equals(version)) {
+					Patch.patch(classpathSorted, nestedJarFolder, "patch-equinox-4.27");
+					SignedJars.stripIf(classpathSorted, jarName -> jarName.startsWith("org.eclipse.osgi"));
+				}
 			}
 			SignedJars.stripIfNecessary(classpathSorted);
 
@@ -106,6 +116,15 @@ public class BuildPluginIdeMain {
 			var ideHooksCopy = ideHooks.copy();
 			ideHooksCopy.add(IdeHookLockFile.forWorkspaceDirAndClasspath(workspaceDir, classpathSorted));
 			SerializableMisc.toFile(ideHooksCopy, ideHooksFile);
+
+			var installDir = workspaceDir.toPath().resolve("install");
+			Files.createDirectories(installDir);
+			var bundlesInfo =
+					workspaceDir
+							.toPath()
+							.resolve("config/org.eclipse.equinox.simpleconfigurator/bundles.info");
+			Files.createDirectories(bundlesInfo.getParent());
+			Files.writeString(bundlesInfo, bundlesDotInfo(classpathSorted));
 
 			debugClasspath.printWithHead(
 					"jars about to be launched", classpathSorted.stream().map(File::getAbsolutePath));
@@ -226,6 +245,42 @@ public class BuildPluginIdeMain {
 		}
 	}
 
+	private static String bundlesDotInfo(List<File> cp) {
+		var buffer = new StringBuilder();
+		var newline = "\n";
+		// for a "real" file these should be different in different places...
+		var startLevel = "0";
+		var markedAsStarted = "false";
+
+		buffer.append("#encoding=UTF-8");
+		buffer.append(newline);
+		buffer.append("#version=1");
+		buffer.append(newline);
+		for (var file : cp) {
+			try {
+				SolsticeManifest manifest = SolsticeManifest.parseJar(file);
+				if (manifest == null
+						|| manifest.getSymbolicName() == null
+						|| manifest.getVersion() == null) {
+					continue;
+				}
+				buffer.append(manifest.getSymbolicName());
+				buffer.append(',');
+				buffer.append(manifest.getVersion());
+				buffer.append(',');
+				buffer.append(file.toURI());
+				buffer.append(',');
+				buffer.append(startLevel);
+				buffer.append(',');
+				buffer.append(markedAsStarted);
+				buffer.append(newline);
+			} catch (Exception e) {
+				// do nothing
+			}
+		}
+		return buffer.toString();
+	}
+
 	private static <T> T parseArg(
 			String[] args, String arg, Function<String, T> parser, T defaultValue) {
 		for (int i = 0; i < args.length - 1; ++i) {
@@ -276,6 +331,9 @@ public class BuildPluginIdeMain {
 		}
 
 		var props = new LinkedHashMap<String, String>();
+		props.put(
+				"org.eclipse.equinox.simpleconfigurator.configUrl",
+				"file: broken on purpose to disable simpleconfigurator");
 		props.put("gosh.args", "--quiet --noshutdown");
 		props.put("osgi.nl", "en_US");
 		props.put("eclipse.noRegistryFlushing", "true");
@@ -296,6 +354,7 @@ public class BuildPluginIdeMain {
 			solstice.openShim(props);
 			ShimIdeBootstrapServices.apply(props, solstice.getContext());
 		}
+		ShimIdeBootstrapServices.shimAndAtomos(props, solstice.getContext());
 
 		solstice.start("org.apache.felix.scr");
 		solstice.startAllWithLazy(false);
@@ -304,7 +363,7 @@ public class BuildPluginIdeMain {
 			// the spelled-out package is on purpose so that Atomos can remain an optional component
 			// works together with
 			// https://github.com/equodev/equo-ide/blob/aa7d30cba9988bc740ff4bc4b3015475d30d187c/solstice/build.gradle#L16-L22
-			dev.equo.solstice.BundleContextAtomos.urlWorkaround(solstice.getContext());
+			dev.equo.solstice.BundleContextAtomos.urlWorkaround(solstice);
 		}
 		if (!initOnly) {
 			ideHooks.forEach(IdeHookInstantiated::afterOsgi, solstice.getContext());